% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textRazor.R
\name{count_char}
\alias{count_char}
\title{Number of characters: function count_char()
This function counts the number of characters (including spaces) in the given text.}
\usage{
count_char(text)
}
\arguments{
\item{text}{A character vector or string containing the text to be counted.}
}
\value{
The number of characters in the text.
}
\description{
Number of characters: function count_char()
This function counts the number of characters (including spaces) in the given text.
}
\examples{
count_characters("Hello, how are you?") # Output: 19
count_characters("R is an amazing programming language.") # Output: 37

}
