% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textRazor.R
\name{count_spaces}
\alias{count_spaces}
\title{Count Spaces}
\usage{
count_spaces(text)
}
\arguments{
\item{text}{A character vector containing the input text.}
}
\value{
An integer representing the number of spaces in the text.
}
\description{
This function takes a text as input and returns the number of spaces in the text.
}
\examples{
text <- "This is a simple example. This is a simple text."
count_spaces(text) output: 9

}
