% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textRazor.R
\name{count_specs}
\alias{count_specs}
\title{Count Special Characters: count_specs()}
\usage{
count_specs(text)
}
\arguments{
\item{text}{A character vector containing the input text.}
}
\value{
An integer representing the number of special characters in the text.
}
\description{
This function takes a text as input and returns the number of special characters in the text.
Special characters are any non-alphabetic and non-numeric characters.
}
\examples{
text <- "?? This is a simple example!! \%\% This is a simple text.&$"
count_specs(text) output: 20

}
