% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textRazor.R
\name{count_words}
\alias{count_words}
\title{Functions for textRazor Package
Number of Words: function count_words()
This function counts the number of words in the given text.}
\usage{
count_words(text)
}
\arguments{
\item{text}{A character vector or string containing the text to be counted.}
}
\value{
The number of words in the text.
}
\description{
Functions for textRazor Package
Number of Words: function count_words()
This function counts the number of words in the given text.
}
\examples{
count_words("Hello, how are you?") # Output: 4
count_words("textRazor is a powerful R package designed for efficient
and precise text analysis. With user-friendly functions,
it provides features like word counting, character counting,
and space counting. Seamlessly handle text data and gain
valuable insights with textRazor, your ultimate tool for
text exploration and understanding.") # Output: 45

}
