% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textRazor.R
\name{word_freq}
\alias{word_freq}
\title{Word Frequency: function word_freq()}
\usage{
word_freq(text)
}
\arguments{
\item{text}{A character vector containing the input text.}
}
\value{
A data frame with two columns: "Word" and "Frequency".
}
\description{
This function takes a text as input and returns a data frame with word frequencies.
It counts the frequency of each word in the text.
}
\examples{
word_freq("This is a simple example. This is a simple text.")
output:      Word Frequency
      1       a         2
      2      is         2
      3  simple         2
      4    this         2
      5 example         1
      6    text         1

}
